+++ incl/MIS_Rank.h
@@ -7,7 +7,7 @@
 
 struct ROW {
 	unsigned long long Player;
-	unsigned long Time;
+	double Time;
 	std::string Name;
 };
 
@@ -20,7 +20,7 @@ public:
 
 	void SaveList();
 	void SortList();
-	int AddPlayer(unsigned long long PMVPlayerID, unsigned long Playtime, std::string &sRankName);
+	int AddPlayer(unsigned long long PMVPlayerID, double Playtime, std::string &sRankName);
 	int RankMode;
 
 	std::vector<ROW*> RankRows;
